Adding a channel buffer to validators
This change introduces a buffer of 100 messages to each producer queue.
Since the kafka client only flushes the internal batch queue when 100
messages have been buffered OR the batch is 1 second old AND we are
currently using a synchronous channel the effect is most calls to
validation wait up to one second to send.

This causes, possibly many, goroutines to stack up in memory waiting for
their turn to return from Validate and be collected.

Signed-off-by: Jesse Jaggars <jjaggars@redhat.com>