feat: Support legacy test connection
The old clients send a test payload in the body of the request which we
weren't checking for. This was resulting in 400 error codes particularly
for satellite clients.
This feature support the JSON body implementation of the test connection
feature in the client.

JIRA: https://issues.redhat.com/browse/RHCLOUD-11259

Signed-off-by: Stephen Adams <tsadams@gmail.com>