[feat] Add payload tracking endpoin
This allows users to hit a /track endpoint on ingress to check where
their payload is in the process of beign ingested into the platform.
Currently it will just return the JSON provided by payload tracker
directly. It will be up to the end user or potentially the client to
pick out the needed data.

This is a first pass, so there's plenty of room for improvement as we
move forward.

RHCLOUD-14381

Signed-off-by: Stephen Adams <tsadams@gmail.com>