[feat] embed openapi spec
Rather than reading from the filesystem every time we want to present
the api spec. We can embed the spec in the executable.

This required moving api into a path where main.go is. The 'embed' lib
cannnot read relative paths so it makes it necessary to put it where the
path will be consistent

RHCLOUD-16031

Signed-off-by: Stephen Adams <tsadams@gmail.com>