[feat] Create single topic for new uploads
New uploads can be advertised on a single topic and the header can be
used if it is meant for a downstream service. This is meant to simplify
ingress code in the long run. It also allows services to onboard easier
by just sending data into the platform for it to be sorted at the client
level.

Signed-off-by: Stephen Adams <tsadams@gmail.com>